Rename main modules in nlp workers to 'xyz_wrapper' to better indicate what they do.
