+++ test/unit/worker/nlp/test_annotate.py
@@ -5,7 +5,7 @@ from unittest.mock import patch
 import cassis
 
 from test.unit.worker.nlp.text_analyzer_mock import TextAnalyzer as MockAnalyzer, MockDAIEntity
-from workers.nlp.annotate.annotate import annotate_text, annotate_xmi
+from workers.nlp.annotate.nlp_components_wrapper import annotate_text, annotate_xmi
 
 _example_xmi_with_pages = """<?xml version='1.0' encoding='ASCII'?>
 <xmi:XMI xmlns:xmi="http://www.omg.org/XMI" xmlns:cas="http:///uima/cas.ecore" xmlns:LayoutElement="http:///org/dainst/nlp/LayoutElement.ecore" xmi:version="2.0">
@@ -35,7 +35,7 @@ class AssertsXmiCanBeLoadedWithDaiTypesystem:
             raise AssertionError(e, "Loading the annotated xmi with our typesystem failed.")
 
 
-@patch('workers.nlp.annotate.annotate._init_text_analyzer')
+@patch('workers.nlp.annotate.nlp_components_wrapper._init_text_analyzer')
 class NlpComponentsAnnotationTest(unittest.TestCase, AssertsXmiCanBeLoadedWithDaiTypesystem):
 
     @staticmethod
